NF: move LoadDeckCounts where its used
